Rename clientsresolver
