+++ libs/api/domains/auth-admin/src/lib/auth-admin.module.ts
@@ -5,7 +5,7 @@ import { AuthAdminApiClientModule } from '@island.is/clients/auth/admin-api'
 import { TenantResolver } from './tenant/tenant.resolver'
 import { TenantEnvironmentResolver } from './tenant/tenant-environment.resolver'
 import { TenantsService } from './tenant/tenants.service'
-import { ClientsResolver } from './client/clientsResolver'
+import { ClientsResolver } from './client/clients.resolver'
 import { ClientsService } from './client/clients.service'
 
 @Module({
